Simplify sqs.
Removes support for sns and such. Now that backend is more simple sqs can be too.
