--- lib/qu-aws.rb
@@ -1,4 +1,0 @@
-require 'qu'
-require 'qu/backend/aws'
-
-Qu.backend = Qu::Backend::AWS.new
