+++ lib/qu-sqs.rb
@@ -1,0 +1,4 @@
+require 'qu'
+require 'qu/backend/sqs'
+
+Qu.backend = Qu::Backend::SQS.new
